requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
dj-pagination==2.4.0
Django==2.2.4
idna==2.8
importlib-metadata==0.20
kombu==4.6.4
more-itertools==7.2.0
pymssql==2.1.4
PyMySQL==0.9.3
pyodbc==4.0.27
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
uuid==1.30
vine==1.3.0
zipp==0.6.0
